requirements.txt
discord.py
ollama
fastf1
requests
discord.py[voice]